+++ sdk/resourcemanager/relay/armrelay/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 0.2.0 (2021-10-28)
+### Breaking Changes
+
+- Function `NewOperationsClient` parameter(s) have been changed from `(*arm.Connection)` to `(azcore.TokenCredential, *arm.ClientOptions)`
+- Function `NewNamespacesClient` parameter(s) have been changed from `(*arm.Connection, string)` to `(string, azcore.TokenCredential, *arm.ClientOptions)`
+- Function `NewHybridConnectionsClient` parameter(s) have been changed from `(*arm.Connection, string)` to `(string, azcore.TokenCredential, *arm.ClientOptions)`
+- Function `NewWCFRelaysClient` parameter(s) have been changed from `(*arm.Connection, string)` to `(string, azcore.TokenCredential, *arm.ClientOptions)`
+
+### New Content
+
+
+Total 4 breaking change(s), 0 additive change(s).
+
+
 ## 0.1.1 (Unreleased)
 
 ### Features Added